+++ CHANGELOG.md
@@ -14,3 +14,4 @@
     - add department
     - add department role
     - 31-12-2024 to do for tommorrow 1 jan 2k25 to fix Offering 
+    - update offering and offering category for crud and add chart